requirements.txt
coverage==4.4.1
Django==1.11.4
django-extensions==1.9.0
django-nested-admin==3.0.20
django-registration==2.3
ipython==6.1.0
pylint==1.7.4
pylint-django==0.7.2
